+++ pkg/deployment/reconcile/plan_builder_maintenance.go
@@ -22,8 +22,11 @@ package reconcile
 
 import (
 	"context"
+	"time"
 
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
+	"github.com/arangodb/kube-arangodb/pkg/deployment/actions"
+	"github.com/arangodb/kube-arangodb/pkg/deployment/features"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
 )
 
@@ -121,4 +124,58 @@ func (r *Reconciler) createMaintenanceConditionPlan(ctx context.Context, apiObje
 
 		return nil
 	}
+}
+
+func (r *Reconciler) createMaintenanceManagementPlan(ctx context.Context, apiObject k8sutil.APIObject,
+	spec api.DeploymentSpec, status api.DeploymentStatus,
+	planCtx PlanBuilderContext) api.Plan {
+	if spec.Mode.Get() == api.DeploymentModeSingle {
+		return nil
+	}
+
+	if !features.Maintenance().Enabled() {
+		// Maintenance feature is not enabled
+		return nil
+	}
+
+	agencyState, agencyOK := planCtx.GetAgencyCache()
+	if !agencyOK {
+		r.log.Error("Unable to get agency mode")
+		return nil
+	}
+
+	if agencyState.Target.HotBackup.Create.Exists() {
+		r.log.Info("HotBackup in progress")
+		return nil
+	}
+
+	enabled := agencyState.Supervision.Maintenance.Exists()
+	c, cok := status.Conditions.Get(api.ConditionTypeMaintenance)
+
+	if (cok && c.IsTrue()) != enabled {
+		// Condition not yet propagated
+		r.log.Info("Condition not yet propagated")
+		return nil
+	}
+
+	if cok {
+		if t := c.LastTransitionTime.Time; !t.IsZero() {
+			if time.Since(t) < 5*time.Second {
+				// Did not elapse 5 s
+				return nil
+			}
+		}
+	}
+
+	if !enabled && spec.Database.GetMaintenance() {
+		r.log.Info("Enabling maintenance mode")
+		return api.Plan{actions.NewClusterAction(api.ActionTypeEnableMaintenance)}
+	}
+
+	if enabled && !spec.Database.GetMaintenance() {
+		r.log.Info("Disabling maintenance mode")
+		return api.Plan{actions.NewClusterAction(api.ActionTypeDisableMaintenance)}
+	}
+
+	return nil
 }